requirements.txt
bottle==0.12.8
django-channels==0.3.0
requests==2.7.0
